ssh_utils: Don't generate new keys from the no_pass_disable script
The purpose of this function is to remove a public key from the remote
authorized_keys file in order to disable no-password login.  It doesn't seem to
make sense to generate a new key if it is missing.  Instead the function should
do nothing if the key file cannot be found.

The generation of new keys is likely a mistake from copying the "enable"
script (where it makes sense to generate  a new key if it does not exist). I
would presume no one is depending on the previous behavior and therefore changed
the script to end the function if the key is missing.

It appears that previously script would have continued the function to try to
remove the newly generated key (which by definition cannot exist in the
authorized_keys file).